pyproject.toml
[tool.poetry]
name = "gas-balance-risk"
version = "0.1.0"
description = ""
authors = ["Moses McCall <mosesmc52@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
scrapy = "^2.13.4"
scrapy-splash = "^0.11.1"
scrapy-user-agents = "^0.1.1"
pandas = "^2.3.3"
eia-ng-client = "^0.1.1"
streamlit = "^1.52.2"
pymongo = "^4.15.5"


[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
